Separate out iterator tests and slightly expand them
